requirements.txt
apprise==0.9.2
configparser==5.0.2
eventlet==0.30.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
pipenv==2021.5.29
protobuf==3.17.3
pycares==3.1.1
pycoingecko==2.2.0
pylint==2.9.6
pymongo==3.12.0
pyOpenSSL==20.0.1
python-telegram-bot==13.7
schedule==1.1.0
Twisted==21.7.0
wcwidth==0.2.5
wheel==0.37.0
tweepy=4.1.0
numpy=1.21.2 
pandas=1.3.3
nltk=3.6.5
textblob=0.15.3
arrow=1.2.0
sdnotify=0.3.2
rapidjson=1.0.0
switch=1.1.0